refactor(test): consolidate shared fixtures into chess-corners-testutil
Extract the synthetic-board generators (the AA board alone had 8 copies), gaussian_blur, add_gaussian_noise and corner-count helpers into a new zero-dependency, publish=false chess-corners-testutil dev-dependency crate so every crate's tests, benches and examples share one byte-identical source. Add a ClassicRefiners trait + RefineSampleSink to replace the copy-pasted refiner-dispatch bodies. Net -1049 LOC on tracked files; fixtures are unchanged, so behavior and the perf baseline are preserved.
