+++ crates/chess-corners-core/Cargo.toml
@@ -35,6 +35,9 @@ log.workspace = true
 serde = { workspace = true, default-features = false, features = ["derive"] }
 
 [dev-dependencies]
+# Path-only (no version) so it is stripped from the published manifest;
+# `chess-corners-testutil` is an unpublished workspace dev-utility.
+chess-corners-testutil = { path = "../chess-corners-testutil" }
 criterion.workspace = true
 image.workspace = true
 serde_json.workspace = true
