+++ crates/chess-corners-core/benches/nms_scaling.rs
@@ -15,11 +15,9 @@
 //! `unstable` module (the curated home for bench-facing sub-stage entry
 //! points); it carries no semver guarantee.
 
-mod common;
-
 use chess_corners_core::chess_response_u8;
 use chess_corners_core::unstable::{detect_peaks_from_response_with_refine_radius, ChessParams};
-use common::synth_chessboard;
+use chess_corners_testutil::synth_chessboard;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
 use std::hint::black_box;
