+++ crates/chess-corners/examples/bench_sweep.rs
@@ -33,111 +33,14 @@
 //!     > book/src/img/bench/bench_sweep.json
 //! ```
 
-use std::time::Instant;
-
 use chess_corners::low_level::ImageView;
+use chess_corners_core::chess_response_u8;
 use chess_corners_core::unstable::ChessParams;
-use chess_corners_core::{
-    chess_response_u8, CenterOfMassConfig, CenterOfMassRefiner, CornerRefiner, ForstnerConfig,
-    ForstnerRefiner, RefineContext, RefineStatus, ResponseMap, SaddlePointConfig,
-    SaddlePointRefiner,
-};
+use chess_corners_testutil::{aa_chessboard, add_gaussian_noise, gaussian_blur};
 
-// ---------------------------------------------------------------------------
-// Synthetic fixture — identical to `refiner_benchmark.rs`.
-
-fn synthetic_chessboard_aa(
-    size: usize,
-    cell: usize,
-    offset: (f32, f32),
-    dark: u8,
-    bright: u8,
-) -> Vec<u8> {
-    const SUPER: usize = 8;
-    let (ox, oy) = offset;
-    let c = cell as f32;
-    let dark_f = dark as f32;
-    let bright_f = bright as f32;
-    let inv_super2 = 1.0 / (SUPER * SUPER) as f32;
-    let mut img = vec![0u8; size * size];
-    for y in 0..size {
-        for x in 0..size {
-            let mut acc = 0.0f32;
-            for sy in 0..SUPER {
-                let yf = y as f32 + (sy as f32 + 0.5) / SUPER as f32 - 0.5;
-                let cy = ((yf - oy) / c).floor() as i32;
-                for sx in 0..SUPER {
-                    let xf = x as f32 + (sx as f32 + 0.5) / SUPER as f32 - 0.5;
-                    let cx = ((xf - ox) / c).floor() as i32;
-                    let dark_cell = (cx + cy).rem_euclid(2) == 0;
-                    acc += if dark_cell { dark_f } else { bright_f };
-                }
-            }
-            img[y * size + x] = (acc * inv_super2).round().clamp(0.0, 255.0) as u8;
-        }
-    }
-    img
-}
-
-fn gaussian_blur(img: &mut [u8], size: usize, sigma: f32) {
-    if sigma <= 0.0 {
-        return;
-    }
-    let radius = ((3.0 * sigma).ceil() as usize).max(1);
-    let klen = 2 * radius + 1;
-    let mut kernel = vec![0f32; klen];
-    let mut sum = 0f32;
-    for (i, k) in kernel.iter_mut().enumerate() {
-        let x = i as f32 - radius as f32;
-        *k = (-(x * x) / (2.0 * sigma * sigma)).exp();
-        sum += *k;
-    }
-    for k in kernel.iter_mut() {
-        *k /= sum;
-    }
-    let mut tmp = vec![0f32; size * size];
-    for y in 0..size {
-        for x in 0..size {
-            let mut acc = 0f32;
-            for (ki, &k) in kernel.iter().enumerate() {
-                let sx = (x as i32 + ki as i32 - radius as i32).clamp(0, size as i32 - 1) as usize;
-                acc += img[y * size + sx] as f32 * k;
-            }
-            tmp[y * size + x] = acc;
-        }
-    }
-    for y in 0..size {
-        for x in 0..size {
-            let mut acc = 0f32;
-            for (ki, &k) in kernel.iter().enumerate() {
-                let sy = (y as i32 + ki as i32 - radius as i32).clamp(0, size as i32 - 1) as usize;
-                acc += tmp[sy * size + x] * k;
-            }
-            img[y * size + x] = acc.round().clamp(0.0, 255.0) as u8;
-        }
-    }
-}
-
-fn add_gaussian_noise(img: &mut [u8], sigma: f32, seed: u64) {
-    if sigma <= 0.0 {
-        return;
-    }
-    let mut state = seed ^ 0x9E3779B97F4A7C15;
-    let mut next_u32 = || {
-        state = state
-            .wrapping_mul(6_364_136_223_846_793_005)
-            .wrapping_add(1_442_695_040_888_963_407);
-        (state >> 33) as u32
-    };
-    let mut uniform = || -> f32 { (next_u32() as f32 + 1.0) / (u32::MAX as f32 + 2.0) };
-    for px in img.iter_mut() {
-        let u1 = uniform();
-        let u2 = uniform();
-        let z = (-2.0 * u1.ln()).sqrt() * (2.0 * core::f32::consts::PI * u2).cos();
-        let v = *px as f32 + z * sigma;
-        *px = v.round().clamp(0.0, 255.0) as u8;
-    }
-}
+#[path = "../tests/common/refiner_harness.rs"]
+mod harness;
+use harness::{bench_refiner, ClassicRefiners};
 
 // ---------------------------------------------------------------------------
 // Per-condition stats bundle.
@@ -205,123 +108,14 @@ struct Stats {
 }
 
 // ---------------------------------------------------------------------------
-// Refiner harness.
+// Aggregator wiring: route the shared harness samples into this file's `Bucket`.
 
-struct ClassicRefiners {
-    center: CenterOfMassRefiner,
-    forstner: ForstnerRefiner,
-    saddle: SaddlePointRefiner,
-}
-
-impl ClassicRefiners {
-    fn new() -> Self {
-        Self {
-            center: CenterOfMassRefiner::new(CenterOfMassConfig::default()),
-            forstner: ForstnerRefiner::new(ForstnerConfig::default()),
-            saddle: SaddlePointRefiner::new(SaddlePointConfig::default()),
-        }
+impl harness::RefineSampleSink for Bucket {
+    fn record_accuracy(&mut self, err: Option<f32>) {
+        self.push_err(err);
     }
-}
-
-fn bench_refiner<R: CornerRefiner>(
-    refiner: &mut R,
-    iters: u64,
-    view: ImageView<'_>,
-    response: Option<&ResponseMap>,
-    seed: [f32; 2],
-    truth: (f32, f32),
-    bucket: &mut Bucket,
-) {
-    let ctx = RefineContext::new(Some(view), response);
-    let _ = refiner.refine(seed, ctx);
-    let first = refiner.refine(seed, ctx);
-    let err = if first.status == RefineStatus::Accepted {
-        let dx = first.x - truth.0;
-        let dy = first.y - truth.1;
-        Some((dx * dx + dy * dy).sqrt())
-    } else {
-        None
-    };
-    bucket.push_err(err);
-
-    let start = Instant::now();
-    for _ in 0..iters {
-        let _ = refiner.refine(seed, ctx);
-    }
-    bucket.add_time(iters, start.elapsed().as_nanos());
-}
-
-// ---------------------------------------------------------------------------
-// ML refiner wrapper (feature-gated).
-
-#[cfg(feature = "ml-refiner")]
-mod ml {
-    use super::*;
-    use chess_corners_ml::{MlModel, ModelSource};
-
-    pub struct MlRefiner {
-        pub model: MlModel,
-        patch_size: usize,
-        buffer: Vec<f32>,
-    }
-
-    impl MlRefiner {
-        pub fn load() -> Option<Self> {
-            let model = MlModel::load(ModelSource::EmbeddedDefault).ok()?;
-            let patch_size = model.patch_size();
-            let buffer = vec![0.0f32; patch_size * patch_size];
-            Some(Self {
-                model,
-                patch_size,
-                buffer,
-            })
-        }
-        fn extract(&mut self, view: ImageView<'_>, x: f32, y: f32) -> Option<()> {
-            let ps = self.patch_size;
-            let half = (ps as f32 - 1.0) * 0.5;
-            let (w, h) = (view.width as f32, view.height as f32);
-            if x - half < 0.0 || y - half < 0.0 || x + half > w - 1.0 || y + half > h - 1.0 {
-                return None;
-            }
-            for iy in 0..ps {
-                let gy = y + iy as f32 - half;
-                for ix in 0..ps {
-                    let gx = x + ix as f32 - half;
-                    self.buffer[iy * ps + ix] = view.sample_bilinear(gx, gy) / 255.0;
-                }
-            }
-            Some(())
-        }
-        pub fn refine(&mut self, view: ImageView<'_>, seed: [f32; 2]) -> Option<[f32; 2]> {
-            self.extract(view, seed[0], seed[1])?;
-            let preds = self.model.infer_batch(&self.buffer, 1).ok()?;
-            let pred = preds.first()?;
-            Some([seed[0] + pred[0], seed[1] + pred[1]])
-        }
-    }
-
-    pub fn bench(
-        refiner: &mut MlRefiner,
-        iters: u64,
-        view: ImageView<'_>,
-        seed: [f32; 2],
-        truth: (f32, f32),
-        bucket: &mut Bucket,
-    ) {
-        let _ = refiner.refine(view, seed);
-        let first = refiner.refine(view, seed);
-        let err = first.map(|r| {
-            let dx = r[0] - truth.0;
-            let dy = r[1] - truth.1;
-            (dx * dx + dy * dy).sqrt()
-        });
-        bucket.push_err(err);
-
-        let start = Instant::now();
-        for _ in 0..iters {
-            let _ = refiner.refine(view, seed);
-        }
-        bucket.add_time(iters, start.elapsed().as_nanos());
+    fn record_timing(&mut self, iters: u64, elapsed_ns: u128) {
+        self.add_time(iters, elapsed_ns);
     }
 }
 
@@ -345,7 +139,7 @@ fn run_condition(
     let mut b_forstner = Bucket::default();
     let mut b_saddle = Bucket::default();
     #[cfg(feature = "ml-refiner")]
-    let mut ml_refiner = ml::MlRefiner::load();
+    let mut ml_refiner = harness::ml::MlRefiner::load();
     #[cfg(feature = "ml-refiner")]
     let mut b_ml = Bucket::default();
 
@@ -355,7 +149,7 @@ fn run_condition(
         for ky in 0..N {
             let ox = CENTER + kx as f32 / N as f32;
             let oy = CENTER + ky as f32 / N as f32;
-            let mut img = synthetic_chessboard_aa(SIZE, cell, (ox, oy), 30, 230);
+            let mut img = aa_chessboard(SIZE, cell, (ox, oy), 30, 230);
             gaussian_blur(&mut img, SIZE, blur_sigma);
             add_gaussian_noise(&mut img, noise_sigma, seed_xor ^ ((kx * N + ky) as u64));
             let view = ImageView::from_u8_slice(SIZE, SIZE, &img).unwrap();
@@ -392,7 +186,7 @@ fn run_condition(
 
             #[cfg(feature = "ml-refiner")]
             if let Some(r) = ml_refiner.as_mut() {
-                ml::bench(r, ITERS_SLOW, view, seed, (ox, oy), &mut b_ml);
+                harness::ml::bench(r, ITERS_SLOW, view, seed, (ox, oy), &mut b_ml);
             }
         }
     }
